refactor: 패키지 생성/이동, 메서드명 수정
- 영속화 대상이 되는 객체를 domain -> entity 패키지로 이동
- repository 패키지 생성
- domain 패키지는 비즈니스 규칙을 담는 곳
- 검색 조건이 없는 search() 메서드명을 findAll()로 수정, 다른 메서드명도 로직에 맞게 수정
