+++ src/main/java/roomescape/domain/ReservationTimeStatus.java
@@ -1,5 +1,7 @@
 package roomescape.domain;
 
+import roomescape.entity.ReservationTime;
+
 public class ReservationTimeStatus {
 
     private final ReservationTime reservationTime;
