+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -2,8 +2,8 @@ package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeStatus;
+import roomescape.entity.ReservationTime;
 
 public record ReservationTimeResponse(
         Long id,
