+++ src/main/java/roomescape/entity/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.entity;
 
 public class Theme {
