+++ src/main/java/roomescape/infra/JdbcThemeRepository.java
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme;
-import roomescape.domain.ThemeRepository;
+import roomescape.entity.Theme;
+import roomescape.repository.ThemeRepository;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
