+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.repository;
 
 import java.util.List;
 import java.util.Optional;
+import roomescape.entity.ReservationTime;
 
 public interface ReservationTimeRepository {
