+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -1,8 +1,9 @@
-package roomescape.domain;
+package roomescape.repository;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import roomescape.entity.Theme;
 
 public interface ThemeRepository {
